Add .dts overlay for PRU Cape
I took the pinmux modes from the .dtsi files provided by TI and compiled
them into a .dts file that should handle device configuration and pinmuxing
for the PRU Cape on Linux. Compile and load it the same way that you would any other
device tree overlay.

One caveat: I tried this on the stock 4.1 kernel Ubuntu image (with the latest 3.8.13 kernel
packages installed), and the pins wouldn't mux correctly. However, I tried it with a custom
3.8.13 kernel that I had compiled some time ago, and it worked if I copied over the associated
boot dtb files. It's possible that something has changed with the DeviceTree files that I'm not
aware of. Your results when using this may vary.